style: n월의 총 좋아요 개수 조회 쿼리 후보 중 결정 및 나머지 코드 제거

+++ src/main/java/com/team4/planit/report/ReportService.java
@@ -24,9 +24,7 @@ public class ReportService {
     public ResponseEntity<?> getReport(String month, HttpServletRequest request) {
         Member member = check.validateMember(request);
         List<String> categoryRank = categoryRepositorySupport.findAllCategoryRank(member, month);
-//        Long monthlyTotalLikes = likesRepositorySupport.findMonthlyTotalLikes(member, month);
-//        Integer monthlyTotalLikes = likesRepositorySupport.findMonthlyTotalLikes(member, month);
-        Integer monthlyTotalLikes = Math.toIntExact(likesRepositorySupport.findMonthlyTotalLikes(member, month));
+        Integer monthlyTotalLikes = likesRepositorySupport.findMonthlyTotalLikes(member, month);
 
         return new ResponseEntity<>(Message.success(monthlyTotalLikes), HttpStatus.OK);
     }
